+++ tiger-common/src/main/java/de/gematik/test/tiger/common/data/config/tigerproxy/TigerRoute.java
@@ -7,6 +7,7 @@ package de.gematik.test.tiger.common.data.config.tigerproxy;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import java.io.Serializable;
+import java.util.List;
 import lombok.*;
 
 @NoArgsConstructor
@@ -22,4 +23,5 @@ public class TigerRoute implements Serializable {
   @Builder.Default private boolean internalRoute = false;
   private boolean disableRbelLogging;
   private TigerBasicAuthConfiguration basicAuth;
+  private List<String> criterions;
 }